chore(Providers): move providers to a directory Provider at the root of the source
