+++ src/Process/FeatureProcessing.js
@@ -88,7 +88,7 @@ export default {
         };
 
         context.scheduler.execute(command).then((result) => {
-            // if request return empty json, WFS_Provider.getFeatures return undefined
+            // if request return empty json, WFSProvider.getFeatures return undefined
             if (result) {
                 // call onMeshCreated callback if needed
                 if (layer.onMeshCreated) {
